Make push notifs demo contact Expo service directly
The intermediate Rails server sometimes doesn't send us notifications (maybe because it sends a notif on only the first push? or it sometimes goes down?). Just make NCL contact the Expo server directly, which you never want to do in a real project, but for NCL we want an E2E echo.

Tested by sending notification to self. Also added the `title` field to the notif, can see it on iOS 11 (and maybe earlier versions).

fbshipit-source-id: af8705c